Moving whitewipe to be a mask with modulation hookup
